requirements.txt
imaplib2 
secure-smtplib
python-dotenv
flask
flask-cors
scikit-learn
joblib
gunicorn
APScheduler